+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
 ## 1.0.0-beta.6 - [UNRELEASED]
+### Changed
+- item.data.caspar.data replaces item.data.templateData for structured template data
+### Added
+- The `item.apply` event
 
 ## 1.0.0-beta.5
 ### Changed